refactor: 프로젝트 삭제 확인 및 삭제 로직 분리 (#174)
